bot/requirements.txt
telethon==1.36.0
MetaTrader5==5.0.4424
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
pydantic==2.9.2
pydantic-settings==2.5.2
httpx==0.27.2
aiofiles==24.1.0
loguru==0.7.2
